requirements.txt
pandas
yfinance
timedelta
datetime
streamlit
numpy
pytz
plotly.express
plotly.graph_objects
